refactor(cli): extract source-resolution setup from ProgramData construction (#9412)
Move per_file_ts7016_diagnostics and file_is_esm_map computation into
SourceResolutionSetup/prepare_source_resolution_setup so ProgramContext
construction in collect_diagnostics_with_source_resolutions reads as
orchestration rather than inlining resolution-adjacent setup logic.
